Setup: Project Layout
Created a basic project layout, with amd64 specific code lying within
the arch/amd64 directory.

The Makefile and linker.ld allow for the kernel to be built and linked
properly. The .gitignore specifies git to ignore the intermediate build
output, the final kernel executable, and the CD ISO build output.