Multiboot: Extract boot information
Extracted the memory map, kernel module information, and kernel command
line arguments from the multiboot_info struct passed by GRUB.

The Multiboot-specific information is packaged into a new structure,
allowing the rest of the kernel to operate independently of how it was
initially booted.

This struct, boot_info, is passed to kernel_main(), which will
eventually call the appropriate functions to process the information.