FPU: Add SSE support
Added SSE initialize code, which indicates to the processor that the
kernel supports FPU exception handling and FPU state saving and
restoration.

The FPU state save/restore will be implemented with interrupt handling,
since that would be when the 'FXSAVE' and 'FXRSTOR' instructions would
need to be executed.

libk/string.c contains certain helpful string operations that may be
needed by kernel routines.