Build: Switch to CMake build system
The top-level CMakeLists.txt outlines the "high-level" build, which
would be applicable regardless of what specific target is chosen.

For each supported architecture, there will be a CMakeLists.txt, which
will add arch-specific command line flags and specify which files need
to be built, based on the provided option flags.

The switch from custom Makefile to CMake was necessary, since the
previous system was not robust and would not allow for the required
level of granularity for feature inclusion/exclusion.