x86_64: Add spinlock implementation
Added a simple assembly spinlock implementation using lock cmpxchg. The
assembly functions are exposed via arch/x86_64/include/lock/spinlock.h
which also defines the 'spinlock_t' type, which is the size of a cache
line and has the proper alignment.

The spinlock does not disable or enable interrupts: if the caller
requires this condition, then it is the caller's responsibility to
ensure that is the case.