x86_64: Organize zone-based page frame allocation
Created a 'zone' structure which contains the memory map pointer as well
as the start and end indices of the zone (effectively an array slice).

The 'zone' struct also contains a spinlock, which is acquired and
released in zone_get_chunk() to allow for multiprocessor safety.

The previous commit introduced four different physical memory zones;
however, the "PHYS_NORMAL" zone was eliminated in this commit. The
subsystem should use PHYS_LOW_MEM or PHYS_HIGH_MEM, whichever is
considered most appropriate.