mm: add malloc-style allocator
Add a malloc-style (free list) allocator to handle allocate and free
requests for memory.

This allocator lies ontop of the virtual memory and physical memory
management subsystems and provides a simple interface for allocating and
freeing memory, as needed.

  vm_alloc(...) - returns a pointer to the allocated block of memory
  vm_free(...) - frees the block of memory

The implementation of vm_free(...) comes in the next commit, but its
declaration has been added in this commit.