x86_64: setup GDT
Setup our own GDT and adjust the segment registers to point to the new
segment descriptors.

The EFI code already loads us in 64-bit mode, so there already is an
existing GDT -- however, it is cleaner for the kernel to use its own.

As a part of the early boot process (see arch/x86_64/boot/boot.S), the
kernel loads its own GDT and then performs an IRETQ to properly set the
CS register (since far jumps are not supported in x86_64). The other
segment registers are updated after the IRETQ is performed.

This commit also contains the start of the IDT handling code. The EFI
also loads its own IDT, but that is of no use to the kernel. The
complete setup code will come in a future commit.