x86_64: load GDT after higher-half transition
The initial implementation loaded the GDT before the higher-half
transition and used the physical address (due to identity mapping) as
the base in the GDTR.

However, when interrupts and exceptions are handled, they reference GDT
entries for segment register switches -- so the GDT needs to be mapped
into the higher-half for this to work. So for simplicity, this should
happen after the jump into the higher-half, so that RIP-relative
addressing takes care of determining the GDT virtual base address.

Fixed a small bug in arch/x86_64/irq/idt.c where the wrong gate type was
specified, which ended up triggering a General Protection Fault (GPF)
whenever an exception was raised.