x86_64: setup handlers for each exception ISR
Create handlers __isrX, where X is the exception number, that are
referenced within the IDT.

These handlers *always* push the exception number onto the stack, which
allows the particular exception to be identified from the single common
handler function which processes all IRQs.

To ensure a consistent stack state, the __isrX handlers will also push
an error code to the stack for all exceptions that don't have an error
code pushed by the CPU. This artificial error code is meaningless and
hard-coded to be 0.

The function _irq_common_base saves the current register state onto the
stack and calls irq_common_handler() for actual processing. Upon return,
it restores the register state and performs an IRETQ to the
pre-interrupt context.