mm: scratch: create common interface
The initial implementation of the scratch space interface was using raw
addresses, which makes it very brittle and cumbersome to use.

The new interface is exposed at the arch-independent level, since a
scratch space region is a useful construct for any architectural OS
implementation. It then calls into the arch-specific functions to
perform the actual scratch space management and mapping/unmapping.