mm: rework page frame allocation
Implement a fixed-frame size stack allocator, which allows physical
frames to be allocated and freed.

The initial implementation of physical memory management was simply to
carve out contiguous chunks and let the VMM layer handle the memory.
While that suffices for all heap-based allocation requests, the ability
to allocate and free frames will be necessary for user-space support.

The VMM subsystem needs to be reworked to use the new physical frame
allocator, because the VMM allocator is currently broken as of this
commit.