irq: rename setup and handler functions
Make slight adjustments to the IRQ interface by changing the top-level
install function name: irq_arch_setup() -> arch_irq_setup()

Similarly, the irq_common_handler() function was renamed, along with the
source file it was contained in. It is not *exactly* a common handler,
but rather a common router for IRQs to their actual handlers.