x86_64: rework boot parameters access
Add dedicated function get_boot_params() to retrieve a pointer to the
boot parameters structure.

The initial implementation used a global variable reference, which is
not an ideal solution: multiple portions of code had extern references.

Instead, the new interfaces declares the boot_params variable as static,
so it cannot be referenced outside. The get_boot_params() function
becomes the only way to access the variable.