memfs: map into address space
The loader passes a pointer to a struct which describes the loaded
memfs, if it exists. In the case where no memfs is loaded, the loader
simply indicates that there is an empty memfs.

When there is a real memfs, the kernel will compute the number of pages
needed to map it into the virtual address space and then does so. It
then records the address start and the size -- which comes in use later.