mm: implement address translation
Added a new top-level function - vm_translate(...) - which takes in a
virtual address and returns the associated physical address, if one
exists, otherwise it returns 0.

This is a useful helper function that will go alongside with the planned
page allocation extension to the VM allocator, since it will allow
kernel subsystems to determine physical addresses of pages that were
received from the VM allocator.