mm: implement donation system
Virtual memory that is allocated manually (outside of the VM allocator)
cannot be reclaimed in any way.

The vm_donate(...) function introduces a new donation system which
allows manually allocated memory to be donated back to the VM allocator,
so that is can be reused.

This way, the memory does not go to waste -- instead, it becomes fully
managed by the allocator subsystem, so it can be allocated and freed
repeatedly, as needed.