mm: alloc: support page allocation requests
Implemented support for a page allocation subcomponent of the greater VM
allocator subsystem.

Different kernel components may have the need to allocate page-aligned
and page-sized blocks, which cannot be satisfied using the general
allocator. Hence, the page allocator subcomponent maintains a free list
of pages which it can hand out and reclaim.

The implementation uses a completely separate freelist implementation
from the general allocator. This is for a couple reasons:

  1. Allocations become fast: having a separate list which holds the
     free pages makes allocations constant time!

  2. Single block size: the complexity in the general allocator scheme
     comes from the fact that blocks are different sizes, which makes it
     necessary to have logic to handle block coalescing and block
     splitting, neither which are needed here.

  3. No space overhead: when a page is allocated, the metadata is not
     preserved, since size information is not needed. Therefore, there
     is no space overhead in this implementation.

The paging subcomponent is initialised as a part of the greater
allocator initialisation, since the external view is that the two
components together make one unified subsystem.