mm: check if mapping exists
Create a top-level function to check whether there exists a mapping
within the paging structure for a given virtual address.

  vm_has_mapping(...) - returns true iff there is a mapping

This functionality already technically exists: checking the result of a
call to vm_translate(...), which is actually how the vm_has_mapping(...)
function is implemented internally.

The main purpose of introducing a whole new function is for clarity and
declaring intent: the caller is checking for existence and not trying to
perform a translation.