hosted: mm: implement physical frame allocation
The hosted architecture uses the standard library to allocate a fixed
number of frames (char arrays) which it stores in a stack.

The arch_phys_allocate_frame(...) function pops the top of the stack and
returns the pointer, throwing an exception if it is out-of-memory.

The arch_phys_free_frame(...) function pushes its frame parameter back
onto the stack, so it can be re-allocated by a future request.